pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "chance-sprite"
version = "0.1.0"
description = "Shadowrun SR5 dice bot"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.3",
  "platformdirs>=4.0",
  "msgspec>=0.18.6"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "dpytest",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
chance_sprite = ["emojis/*"]
